-Wsign-compare conversions.c

+++ tests/conversions.c
@@ -62,7 +62,7 @@ main (int    argc,
   putenv ("BABL_DEBUG_MISSING" "=" "1");
   babl_init ();
   
-  for (int i = 0; i < sizeof (fishes)/sizeof(fishes[0]);i ++)
+  for (size_t i = 0; i < sizeof (fishes)/sizeof(fishes[0]);i ++)
   {
     babl_fish (babl_format (fishes[i].from_format),
                babl_format (fishes[i].to_format));
